home page configuration system de securite

+++ app/Http/Controllers/QuestionController.php
@@ -27,7 +27,11 @@ class QuestionController extends Controller
      */
     public function create()
     {
-        return view('question.create');
+        $idAdmin = Option::all()->first()->user_id;
+        if(auth()->user()->id == $idAdmin ){
+            return view('question.create');
+        }
+        return redirect()->route('allDiagnostic');
     }
 
     /**
@@ -38,17 +42,22 @@ class QuestionController extends Controller
      */
     public function store(Request $request)
     {
-        $data = $request->validate([
-            'question' => 'required|max:255',
-            'categorie' => 'required|max:255',
-        ]);
-        $question = new Question();
-        $question->question = $data['question'];
-        $question->categorie_id = $data['categorie'];
-        $question->user_id = auth()->user()->id;
-        $question->slug = Str::slug(Str::random(10));
-        $question->save();
-        return redirect()->route('question');
+        $idAdmin = Option::all()->first()->user_id;
+        if(auth()->user()->id == $idAdmin ){
+            $data = $request->validate([
+                'question' => 'required|max:255',
+                'categorie' => 'required|max:255',
+            ]);
+            $question = new Question();
+            $question->question = $data['question'];
+            $question->categorie_id = $data['categorie'];
+            $question->user_id = auth()->user()->id;
+            $question->slug = Str::slug(Str::random(10));
+            $question->save();
+            return redirect()->route('allQuestion');
+        }
+        
+        return redirect()->route('allDiagnostic');
     }
 
     /**
@@ -57,10 +66,20 @@ class QuestionController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show($id)
+    public function show($slug)
     {
+        $idAdmin = Option::all()->first()->user_id;
+        if(auth()->user()->id == $idAdmin ){
+            $question = Question::where('slug',$slug)->first();
+            if(isset($question)){
+
+                return view('question.show', compact('question'));
+            }
+            return redirect()->route('allQuestion');
+        }
         
-        return Question::where("id", $id)->get();
+
+        return redirect()->route('allDiagnostic');
     }
 
     /**
@@ -71,13 +90,17 @@ class QuestionController extends Controller
      */
     public function edit($slug)
     {
-        $question = Question::where('slug',$slug)->first();
-        if(isset($question)){
 
-            return view('question.update', compact('question'));
+        $idAdmin = Option::all()->first()->user_id;
+        if(auth()->user()->id == $idAdmin ){
+            $question = Question::where('slug',$slug)->first();
+            if(isset($question)){
+
+                return view('question.update', compact('question'));
+            }
+            return redirect()->route('allQuestion');
         }
-
-        return redirect()->route('question');
+        return redirect()->route('allDiagnostic');
     }
 
     /**
@@ -89,20 +112,20 @@ class QuestionController extends Controller
      */
     public function update(Request $request, $id)
     {
-        
-        $data = Question::find($id);
+        $idAdmin = Option::all()->first()->user_id;
+        if(auth()->user()->id == $idAdmin ){
+            $data = Question::find($id);
 
-        if(isset($data)){
-            $data->question = $request->question;
-            $data->categorie_id = $request->categorie;
-            $data->user_id = auth()->user()->id;
-            $data->slug = Str::slug(Str::random(10));
-            $data->save();
+                if(isset($data)){
+                    $data->question = $request->question;
+                    $data->categorie_id = $request->categorie;
+                    $data->user_id = auth()->user()->id;
+                    $data->slug = Str::slug(Str::random(10));
+                    $data->save();
+                }
+            return redirect()->route('allQuestion');
         }
-       
-        
-
-        return redirect()->route('question');
+        return redirect()->route('allDiagnostic');
     }
 
     /**
@@ -113,13 +136,16 @@ class QuestionController extends Controller
      */
     public function destroy($slug)
     {
+        $idAdmin = Option::all()->first()->user_id;
+        if(auth()->user()->id == $idAdmin ){
+            $data = Question::where('slug', $slug)->first();
+            if(isset($data)){
+                $data->delete();
 
-        $data = Question::where('slug', $slug)->first();
-        if(isset($data)){
-            $data->delete();
-            
+            }
+
+            return redirect()->route('allQuestion');
         }
-
-        return redirect()->route('question');
+        return redirect()->route('allDiagnostic');
     }
 }
